Drop kube import from ca auth
