+++ security/pkg/server/ca/authenticate/cert_authenticator.go
@@ -25,9 +25,7 @@ import (
 )
 
 const (
-	bearerTokenPrefix = "Bearer "
-	authorizationMeta = "authorization"
-	clusterIDMeta     = "clusterid"
+	BearerTokenPrefix = "Bearer "
 
 	ClientCertAuthenticatorType = "ClientCertAuthenticator"
 )
