+++ security/pkg/server/ca/authenticate/cert_authenticator_test.go
@@ -23,7 +23,6 @@ import (
 
 	"golang.org/x/net/context"
 	"google.golang.org/grpc/credentials"
-	"google.golang.org/grpc/metadata"
 	"google.golang.org/grpc/peer"
 
 	"istio.io/istio/security/pkg/pki/util"
@@ -122,70 +121,6 @@ func TestAuthenticate_clientCertAuthenticator(t *testing.T) {
 		if !reflect.DeepEqual(tc.caller, result) {
 			t.Errorf("Case %q: Unexpected authentication result: want %v but got %v",
 				id, tc.caller, result)
-		}
-	}
-}
-
-func TestExtractBearerToken(t *testing.T) {
-	testCases := map[string]struct {
-		metadata                 metadata.MD
-		expectedToken            string
-		extractBearerTokenErrMsg string
-	}{
-		"No metadata": {
-			expectedToken:            "",
-			extractBearerTokenErrMsg: "no metadata is attached",
-		},
-		"No auth header": {
-			metadata: metadata.MD{
-				"random": []string{},
-			},
-			expectedToken:            "",
-			extractBearerTokenErrMsg: "no HTTP authorization header exists",
-		},
-		"No bearer token": {
-			metadata: metadata.MD{
-				"random": []string{},
-				"authorization": []string{
-					"Basic callername",
-				},
-			},
-			expectedToken:            "",
-			extractBearerTokenErrMsg: "no bearer token exists in HTTP authorization header",
-		},
-		"With bearer token": {
-			metadata: metadata.MD{
-				"random": []string{},
-				"authorization": []string{
-					"Basic callername",
-					"Bearer bearer-token",
-				},
-			},
-			expectedToken: "bearer-token",
-		},
-	}
-
-	for id, tc := range testCases {
-		ctx := context.Background()
-		if tc.metadata != nil {
-			ctx = metadata.NewIncomingContext(ctx, tc.metadata)
-		}
-
-		actual, err := extractBearerToken(ctx)
-		if len(tc.extractBearerTokenErrMsg) > 0 {
-			if err == nil {
-				t.Errorf("Case %s: Succeeded. Error expected: %v", id, err)
-			} else if err.Error() != tc.extractBearerTokenErrMsg {
-				t.Errorf("Case %s: Incorrect error message: %s VS %s",
-					id, err.Error(), tc.extractBearerTokenErrMsg)
-			}
-			continue
-		} else if err != nil {
-			t.Fatalf("Case %s: Unexpected Error: %v", id, err)
-		}
-
-		if actual != tc.expectedToken {
-			t.Errorf("Case %q: Unexpected token: want %s but got %s", id, tc.expectedToken, actual)
 		}
 	}
 }
